+++ src/app/category-detail/category-detail.component.html
@@ -1,13 +1,5 @@
-<!--
-<ion-segment *ngIf="settings.showGroupAll"  [(ngModel)]="groupSelected">
-    <ion-segment-button *ngFor="let group of groups" [class.latest-selected]="settings.editMode && group == latestSelected" [value]="group" (click)="groupChanged(group);unselectedItem(group)" (press)="editItem(group)">
-      <ion-icon *ngIf="settings.ShowGroup == 0 || settings.ShowGroup == 2" [name]="group.icon"></ion-icon>
-      {{settings.ShowGroup == 1 || settings.ShowGroup == 2 ? group.name  : ''}}
-    </ion-segment-button>
-  </ion-segment>
--->
 <div *ngIf="settings.showGroupAll">
-  <ion-button *ngFor="let group of groups" size="large" [class.button-selected]="groupedsSelected(group)" [class.latest-button-selected]="settings.editMode && group == latestSelected" (click)="groupChanged(group);" (press)="editItem(group)">
+  <ion-button *ngFor="let group of groups" size="default" [class.button-selected]="groupedsSelected(group)" [class.latest-button-selected]="settings.editMode && group == latestSelected" (click)="groupChanged(group);" (press)="editItem(group)">
     <div>
         <ion-icon *ngIf="settings.ShowGroup == 0 || settings.ShowGroup == 2" [name]="group.icon"></ion-icon>
         {{settings.ShowGroup == 1 || settings.ShowGroup == 2 ? group.name  : ''}}
@@ -18,7 +10,7 @@
 <ion-grid>
   <ion-row>
     <ion-col *ngFor="let action of actions | actionsByGroup:actionGroups; index as i; first as isFirst" [sizeLg]="settings.columnsPerRow || 2" [sizeMd]="settings.columnsPerRow || 3" [sizeSm]="settings.columnsPerRow || 4" [sizeXl]="settings.columnsPerRow || 4" [sizeXs]="settings.columnsPerRow || 6">
-        <ion-card class="welcome-card" [class.selected-action]="action == actionSelected" [class.latest-selected]="settings.editMode && action == latestSelected" (click)="unselectedItem(action)" (press)="editItem(action)">
+        <ion-card class="welcome-card" [class.selected-action]="!settings.editMode && action == actionSelected" [class.latest-selected]="settings.editMode && action == latestSelected" (click)="selectedAction(action);unselectedItem(action)" (press)="editItem(action)">
             <ion-img [src]="action.path" onError="this.src='./assets/images/camera-icon.png';"></ion-img>
             <ion-card-header>
               <ion-card-title text-center>{{action.name}}</ion-card-title>
