+++ src/app/category-detail/category-detail.component.scss
@@ -4,6 +4,15 @@
     color: #333;
   } 
 
+    // color que cambia la imagen apretada
+    .selected-action {
+      background-color: #ccc;
+    } 
+  
+    .selected-action ion-card-content h2{
+      color: black ;
+      font-weight: bold;
+    }
   
   ion-button div {
     display: flex;
