+++ src/app/category-edit/category-edit.component.ts
@@ -43,7 +43,7 @@ export class CategoryEditComponent implements OnInit {
     if (!this.category) {
       this.category = new Category();
       this.category.id = new Date().valueOf(); // genera el id unico
-      this.category.name = 'My new Category';
+      this.category.name = 'Mi nueva categoria';
       this.category.icon = 'add-circle';
       this.order = this.categories.length;
     } else {
