+++ src/app/tab1/tab1.page.scss
@@ -3,15 +3,7 @@
   overflow: hidden;
 }
 
-  // color que cambia la imagen apretada
-  .selected-action {
-    background-color: #ccc;
-  } 
 
-  .selected-action ion-card-content h2{
-    color: black ;
-    font-weight: bold;
-  }
 
  //TODO: repeat in other component
   // color que cambia la imagen apretada
